+++ hbs2-core/test/Main.hs
@@ -5,6 +5,7 @@ import TestActors
 import TestFileLogger
 import TestScheduled
 import TestDerivedKey
+import TestMerkleWalk
 
 import Test.Tasty
 import Test.Tasty.HUnit
@@ -19,6 +20,12 @@ main =
       , testCase "testFileLogger" testFileLogger
       , testCase "testScheduledActions" testScheduled
       , testCase "testDerivedKeys1" testDerivedKeys1
+      , testGroup "merkle walk over a graph somebody else chose"
+          [ testCase "enters each node once when asked about a set"
+              testMerkleWalkBounded
+          , testCase "still repeats a leaf that a tree names twice"
+              testMerkleWalkKeepsDuplicates
+          ]
       ]